% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DNA_contamination.R
\name{check_expr_distribution}
\alias{check_expr_distribution}
\title{Expression distribution}
\usage{
check_expr_distribution(
  counts = NULL,
  normalization = c("DESeq2", "qsmooth", "none"),
  metadata = data.frame(sample_name = vector(mode = "character"), group = vector(mode =
    "character"))
)
}
\arguments{
\item{counts}{A numeric matrix or data frame containing gene expression count
data for an RNA-seq experiment, with row names for gene ID and column names
for sample names. If values are not integers, it will be rounded. You can
use the count table outputted by the \code{\link[=summarize_reads]{summarize_reads()}} function with a GTF
file as annotation.}

\item{normalization}{A character(1) vector, specifying a between-sample
normalization methods: DESeq2's  median of ratios method, smooth quantile
normalization method \code{\link[qsmooth:qsmooth]{qsmooth::qsmooth()}}, or \code{none}.}

\item{metadata}{A data frame including column names: sample_name and group
for sample names and experimental groups of samples, respectively. The
order of the sample name in the rows must match those in the count table
specified by \code{counts}.}
}
\value{
A list of 3 ggplot objects.
\describe{
\item{box_plot}{boxplots showing DESeq2-normalized gene-level count
distribution, on a log scale}
\item{density_plot}{density plots showing DESeq2-normalized gene-level
count distribution, on a log scale}
\item{ecd_plot}{plots showing empricial cumulative distribution of fraction
of genes with CPM greater than or equal to a given CPM
on a log scale}
}
}
\description{
Compare expression distribution by boxplot, density plot and empirical
cumulative distribution plot
}
\examples{
\dontrun{
library(patchwork)
tmp_dir <- tempdir()
## download feaureCounts results
count_url <- paste0("https://www.dropbox.com/scl/fi/lyvh6bsljnqxtq85nnugq/",
           "read_count_summary.RData?rlkey=e0tmpehpxtnr1fdx4fz0h8sa0&dl=1")
download.file(url= count_url,
              destfile = file.path(tmp_dir, "read_count_summary.RData"),
              mode = "wb")
load(file.path(tmp_dir, "read_count_summary.RData"))
metadata <- counts_summary$metadata
metadata$group <- gsub("CD1A\\\\(-\\\\)", "CD1A_N", metadata$group)
metadata$group <- gsub("CD1A\\\\(\\\\+\\\\)", "CD1A_P", metadata$group)

p <- check_expr_distribution(counts = counts_summary$gtf$counts,
                             normalization = "DESeq2",
                             metadata = metadata)
wrap_plots(p, ncol = 1, nrow =3, widths = 6, heights = 8)
}
}
