% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DNA_contamination.R
\name{check_read_distribution}
\alias{check_read_distribution}
\title{Check reads distribution over different types of genomic features}
\usage{
check_read_distribution(
  featurecounts_list = NULL,
  metadata = data.frame(sample_name = vector(mode = "character"), group = vector(mode =
    "character"))
)
}
\arguments{
\item{featurecounts_list}{A list of lists, containing output of the
\code{\link[=summarize_reads]{summarize_reads()}} function.}

\item{metadata}{A data frame with column names: sample_name and group for
sample names and experimental groups for each sample, respectively.}
}
\value{
A list of two elements.
\describe{
\item{p}{A ggplot object, showing percentages of fragments assigned to
different genomic features, such as genic regions, intergenic regions,
exonic regions, intronic regions, rRNA genes, mitochondrial genome,
chloroplast genome (only for plants)}
\item{IR_rate}{a data frame containing the percentages of reads mapping to
intergenic regions, which can be used for the "IR\%" method for correcting
for gDNA contamination}
}
}
\description{
Generate plots showing percentages of fragments assigned to different
genomic features.
}
\examples{
tmp_dir <- tempdir()
## download feaureCounts results
count_url <- paste0(
    "https://www.dropbox.com/scl/fi/lyvh6bsljnqxtq85nnugq/",
    "read_count_summary.RData?rlkey=e0tmpehpxtnr1fdx4fz0h8sa0&dl=1"
)
download.file(
    url = count_url,
    destfile = file.path(tmp_dir, "read_count_summary.RData"),
    mode = "wb"
)
load(file.path(tmp_dir, "read_count_summary.RData"))
metadata <- counts_summary$metadata
counts_summary$metadata <- NULL
p <- check_read_distribution(
    featurecounts_list = counts_summary,
    metadata = metadata
)
p$p
}
