% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpRNAseq-package.R
\docType{package}
\name{CleanUpRNAseq-package}
\alias{CleanUpRNAseq}
\alias{CleanUpRNAseq-package}
\title{CleanUpRNAseq: CleanUpRNAseq detects and corrects for genomic DNA contamination in RNA-seq data}
\description{
RNA-seq data generated by some library preparation methods, such as rRNA-depletion-based method and the SMART-seq method, might be contaminated by genomic DNA (gDNA), if DNase I disgestion is not performed properly during RNA preparation. CleanUpRNAseq is developed to check if RNA-seq data is suffered from gDNA contamination. If so, it can perform correction for gDNA contamination and reduce false discovery rate of differentially expressed genes.
}
\author{
\strong{Maintainer}: Haibo Liu \email{haibo.liu@umassmed.edu} (\href{https://orcid.org/0000-0002-4213-2883}{ORCID})

Authors:
\itemize{
  \item Lihua Julie Zhu \email{Julie.Zhu@umassmed.edu}
  \item Kai Hu \email{kai.hu@umassmed.edu}
}

Other contributors:
\itemize{
  \item Kevin O'Connor \email{Kevin.O'Connor@umassmed.edu} [contributor]
  \item Michelle Kelliher \email{Michelle.Kelliher@umassmed.edu} [contributor]
}

}
\keyword{internal}
