% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUpRNAseq-package.R
\docType{package}
\name{CleanUpRNAseq}
\alias{-package}
\alias{CleanUpRNAseq}
\title{A package for detecting and correcting for DNA contamination in RNA-seq data}
\description{
The CleanUpRNAseq package provides two categories of functions for detecting
DNA contamination and correcting for contamination from RNA-seq count data
}
\section{functions for detecting DNA contamination}{
 check_expr_distribution,
ccheck_expressed_gene_percentage, check_read_distribution,
check_sample_correlation, exploratory_analysis, summarize_reads,
get_feature_saf, granges_to_saf, make_ensdb
}

\section{functions for correcting for DNA contamination}{
 global_correction,
filter_unexpressed_genes, get_unexpressed_spliced_genes, salmon_res
}

