% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_correction.R
\name{correct_stranded_lib}
\alias{correct_stranded_lib}
\title{Correct for gDNA contamination in stranded libraries}
\usage{
correct_stranded_lib(
  metadata = data.frame(salmon_quant_file_strand = vector(mode = "character"),
    salmon_quant_file_reverse_strand = vector(mode = "character"), sample_name =
    vector(mode = "character")),
  ensdb_sqlite = NULL
)
}
\arguments{
\item{metadata}{A data frame with column names: BAM_file, sample_name, and
group. The \emph{salmon_quant_file_strand} and
\emph{salmon_quant_file_reverse_strand} column contains the full paths to
\emph{quant.sf} files for quantitation of gene expression by Salmon
pseudo-alignment with the library strandedness type set to the true and
reverse orientations. See Salmon's description of library type
(\url{https://salmon.readthedocs.io/en/latest/library_type.html})
for details. The \emph{sample_name} columns contain the unique sample labels
for each sample. The order of the three columns doesn't matter.}

\item{ensdb_sqlite}{A character(1), a path to an SQLite database for
an object of the \link[ensembldb:EnsDb-class]{ensembldb::EnsDb}.}
}
\value{
A list of of matrices of gene-level abundance, counts, and length.
See \code{\link[tximport:tximport]{tximport::tximport()}}. The count matrix is corrected for DNA
contamination and rounded into integers.
\describe{
\item{abundance}{A numeric matrix containing corrected abundance (TPM) for
each gene of each sample}
\item{counts}{An \emph{integer} matrix containing \emph{corrected} read count for
each gene of each sample}
\item{length}{A numeric matrix containing length (bp) for each gene of
each sample}
}
}
\description{
Correct for gDNA contamination in stranded libraries based on Salmon
quantitation using the real and opposite library strandedness information
}
\examples{
options(timeout = max(3000, getOption("timeout")))
## using made-up salmon quant.sf files, NOT RUN. For real situation,
## replace the quant.sf files and sample names with real ones.
quant_files <-
    system.file("extdata",
                c("K084CD7PCD1N.quant.sf", "K084CD7PCD1P.quant.sf"),
                package = "CleanUpRNAseq")
metadata <- data.frame(
    salmon_quant_file_strand = quant_files,
    salmon_quant_file_reverse_strand = quant_files,
    sample_name = c("CD1AN_m2_1", "CD1AP_m2_1")
)

# download the EnsDb SQLite database: GRCh38.V110.ensdb.sqlite.zip
tmp_dir <- tempdir()
ensdb_url <- paste0(
    "https://zenodo.org/records/11458839/files/",
    "GRCh38.V110.ensdb.sqlite.zip?download=1"
)
download.file(
    url = ensdb_url,
    destfile = file.path(tmp_dir, "GRCh38.V110.ensdb.sqlite.zip"),
    mode = "wb"
)
unzip(file.path(tmp_dir, "GRCh38.V110.ensdb.sqlite.zip"),
    exdir = tmp_dir
)
hs_ensdb_sqlite <- file.path(tmp_dir, "GRCh38.V110.ensdb.sqlite")

corrected_counts <- correct_stranded_lib(
    metadata = metadata,
    ensdb_sqlite = hs_ensdb_sqlite
)

}
