% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{feature_counts_list}
\alias{feature_counts_list}
\title{GC content and lengths of 2000 intergenic regions}
\format{
"intergenic_region", "intronic_region", "rRNA", "mitochondrion", "gtf",
"chloroplast" (plant only).
For genomic features, gene, exon, intron, rRNA, mitochondrion, and
chloroplast, only the \code{stat} elements of the \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}}
output is stored.
For intergenic region, the \code{stat}, \code{annotation} and \code{counts} elements
are stored; for GTF-based summarization, the \code{stat} and \code{counts} elements
are stored.
}
\usage{
data(feature_counts_list)
}
\description{
GC content and lengths of 2000 human intergenic regions calculated using the
\code{\link[=calc_region_gc]{calc_region_gc()}} function.
}
\keyword{datasets}
