% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_bin_correction.R
\name{gc_bias_correction}
\alias{gc_bias_correction}
\title{Correct DNA contamination considering GC-bias effect}
\usage{
gc_bias_correction(
  salmon_res = NULL,
  gene_gc = NULL,
  intergenic_counts = NULL,
  intergenic_gc = NULL,
  plot = FALSE
)
}
\arguments{
\item{salmon_res}{A list of matrices containing gene-level abundances,
counts, lengths, such as the output of the \code{\link[=salmon_res]{salmon_res()}} function.
For more details, See \code{\link[tximport:tximport]{tximport::tximport()}}.}

\item{gene_gc}{A data frame or matrix with two columns: gc_content in
proportion between 0 and 1, and width in basepair, containing
GC-content and total exon lengths of each gene. An output of the
\code{\link[=calculate_gene_gc]{calculate_gene_gc()}} function.}

\item{intergenic_counts}{A data frame or matrix containing counts assigned to
each intergenic region of each sample, such as the sublist, \emph{counts}  from
sublist named \emph{intergenic_region} of an output from the \code{\link[=summarize_reads]{summarize_reads()}}
function.}

\item{intergenic_gc}{A data frame or matrix with two columns: gc_content in
proportion between 0 and 1, and width in basepair, containing
GC-content and lengths of each intergenic region, such as an output from
the \code{\link[=calculate_region_gc]{calculate_region_gc()}} function.}

\item{plot}{A logical(1) vector, specifying whether to output a panel of
scatter plot showing a bi-variate distribution of GC content and count per
base of intergenic regions in each sample. Default is TRUE.}
}
\value{
A data frame containing corrected count for each gene (row) of each
sample (column).
}
\description{
Correct DNA contamination considering GC-bias effect on fragment
amplification. Intergenic regions are binned based on their GC content
ranging from 0 to 100\%, with a bin size of 5\%. Per gene DNA contamination
is estimated as the product of count per base in a GC content matching bin
of intergenic regions and the total collapsed exons of a gene and is
subtracted away from the gene count matrix.
}
\examples{
options(timeout = max(3000, getOption("timeout")))
tmp_dir <- tempdir()
## download feaureCounts results
count_url <- paste0(
    "https://www.dropbox.com/scl/fi/lyvh6bsljnqxtq85nnugq/",
    "read_count_summary.RData?rlkey=e0tmpehpxtnr1fdx4fz0h8sa0&dl=1"
)
download.file(
    url = count_url,
    destfile = file.path(tmp_dir, "read_count_summary.RData"),
    mode = "wb"
)
load(file.path(tmp_dir, "read_count_summary.RData"))

# download the Salmon quantification results
salmon_url <- paste0(
    "https://drive.google.com/",
    "uc?export=download&id=13vsobXnENoiYOBo-Vf_e_00xbR8ekc4Z"
)
salmon_destfile <- file.path(tmp_dir, "salmon_quant_summary.RData")
download.file(
    url = salmon_url,
    destfile = salmon_destfile,
    mode = "wb"
)

## load the salmon_quant object
load(salmon_destfile)

gene_gc_content_url <-
    paste0(
        "https://www.dropbox.com/scl/fi/taoj18w5t4up22d04xihk/GRCh38.gene.",
        "exon.collapsed.GC.content.RDS?rlkey=zj31ul2slckrn5zm8vq6tgq31&dl=1"
    )
gene_gc_destfile <- file.path(
    tmp_dir,
    "GRCh38.gene.exon.collapsed.GC.content.RDS"
)
download.file(
    url = gene_gc_content_url,
    destfile = gene_gc_destfile,
    mode = "wb"
)

## load the gene_ object
gene_gc <- readRDS(gene_gc_destfile)

intergenic_gc_content_url <-
    paste0(
        "https://www.dropbox.com/scl/fi/ik5z5etbr63nnv43jna8c/GRCh38.",
        "intergenic.GC.content.RDS?rlkey=kxrn6ixdnvhpba9h7w2r265xp&dl=1"
    )
intergenic_gc_destfile <- file.path(
    tmp_dir,
    "GRCh38.intergenic.GC.content.RDS"
)
download.file(
    url = intergenic_gc_content_url,
    destfile = intergenic_gc_destfile,
    mode = "wb"
)

## load the salmon_quant object
intergenic_gc <- readRDS(intergenic_gc_destfile)
gc_bias_corrected_count <-
    gc_bias_correction(
        salmon_res = salmon_quant,
        gene_gc = gene_gc,
        intergenic_counts =
            counts_summary$intergenic_region$counts,
        intergenic_gc = intergenic_gc,
        plot = FALSE
    )
}
