% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_bin_correction.R
\name{gc_bin_contamination}
\alias{gc_bin_contamination}
\title{Correct for DNA contamination in consideratio of GC-bias}
\usage{
gc_bin_contamination(intergenic_counts, intergenic_gc, plot = TRUE)
}
\arguments{
\item{intergenic_counts}{A data frame or matrix containing counts assigned to
each intergenic region of each sample, such as the sublist, \emph{counts}  from
sublist named \emph{intergenic_region} of an output from the \code{\link[=summarize_reads]{summarize_reads()}}
function.}

\item{intergenic_gc}{A data frame or matrix with two columns: gc_content in
proportion between 0 and 1, and width in basepair, containing
GC-content and lengths of each intergenic region, such as an output from
the \code{\link[=calculate_region_gc]{calculate_region_gc()}} function.}

\item{plot}{A logical(1) vector, specifying whether to output a panel of
scatter plot showing a bi-variate distribution of GC content and count per
base of intergenic regions in each sample. Default is TRUE. A loess
regression line is displayed in each subplot.}
}
\value{
A matrix containing estimated DNA contamination in the count per base
of intergenic region binned by GC content, with a bin size of 5\% of each
sample, 20 bins. Column names are sample names, rownames are GC content
bins in the form of '(0,0.05]'.
}
\description{
Correct for DNA contamination in consideratio of GC-bias
}
