% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_GC}
\alias{gene_GC}
\title{GC content and lengths of 2000 human genes}
\format{
A data frame with 2000 rows and 2 columns:
\describe{
\item{gc_content}{metagene-level GC content, values are between 0 and 1}
\item{width}{length of metagenes in base pair(bp)}
}
}
\source{
Homo_sapiens.GRCh38.110.gtf.gz
}
\usage{
data(gene_GC)
}
\description{
GC content and lengths of 2000 human Ensembl genes. For each gene exons
are collapsed to get disjoint meta-exons and GC content is calculated
for each meta-exon. A exon length-weighted GC content is calculated for
each gene using the \code{\link[=calc_gene_gc]{calc_gene_gc()}} function.
}
\keyword{datasets}
