% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DNA_contamination.R
\name{granges_to_saf}
\alias{granges_to_saf}
\title{Convert genomic features from GRanges to Simplified Annotation Format (SAF)}
\usage{
granges_to_saf(granges)
}
\arguments{
\item{granges}{An object of \link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}}
}
\value{
A data frame with columns: "GeneID", "Chr", "Start", "End","Strand".
}
\description{
Convert genomic features from GRanges to Simplified Annotation Format (SAF)
}
\examples{
# example code
library("GenomicRanges")
gr0 <- GRanges(Rle(c("chr2", "chr2", "chr1", "chr3"),
               c(1, 3, 2, 4)), IRanges(1:10, width=10:1))
saf <- granges_to_saf(gr0)

}
