% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intergenic_GC}
\alias{intergenic_GC}
\title{GC content and lengths of 2000 intergenic regions}
\format{
A data frame with 2000 rows and 2 columns:
\describe{
\item{gc_content}{metagene-level GC content, values are between 0 and 1}
\item{width}{length of metagenes in base pair(bp)}
}
}
\source{
Homo_sapiens.GRCh38.110.gtf.gz
}
\usage{
data(intergenic_GC)
}
\description{
GC content and lengths of 2000 human intergenic regions calculated using the
\code{\link[=calc_region_gc]{calc_region_gc()}} function.
}
\keyword{datasets}
