% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ensdb.R
\name{make_ensdb}
\alias{make_ensdb}
\title{Make an EnsDb from a GTF file}
\usage{
make_ensdb(
  gtf = NULL,
  outfile = NULL,
  organism_latin_name = "Homo_sapiens",
  genome_version = "GRCh38",
  Ensembl_release_version = 110
)
}
\arguments{
\item{gtf}{A character(1), a path to an Ensembl GTF file.}

\item{outfile}{A character(1), a path to a SQLite file to store an object of
the \link[ensembldb:EnsDb-class]{ensembldb::EnsDb}.}

\item{organism_latin_name}{A character(1), Latin name for the organism, with
spaces replaced by underscores "_".}

\item{genome_version}{A character(1), the assembly version of the reference
genome, such as "GRCh38" for the human reference genome version 38.}

\item{Ensembl_release_version}{An integer(1), the Ensembl release version of
the reference genome with patches, such as 110 for Ensembl Release 110
(July 2023).}
}
\value{
A name of an SQLite file for the built EnsDb. Use the
\code{\link[ensembldb:EnsDb]{ensembldb::EnsDb()}} function from the package ensembldb package to
load the SQLite file to the R environment.
}
\description{
A wrapper function from the ensembldb package to make an EnsDb SQLite
database from an Ensembl GTF file.
}
\examples{
gtf <- system.file("extdata", "example.gtf.gz",
                   package = "CleanUpRNAseq")
tmp_dir <- tempdir()
hs_ensdb_sqlite <- make_ensdb(
    gtf = gtf,
    outfile = file.path(tmp_dir, "EnsDb.hs.v110.sqlite"),
    organism_latin_name = "Homo_Sapiens",
    genome_version = "GRCh38",
    Ensembl_release_version = 110
)
}
