% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DNA_contamination.R
\name{retry_download}
\alias{retry_download}
\title{Download files via HTTP with retry if failure occurs}
\usage{
retry_download(
  expr,
  isError = function(x) {
     "try-error" \%in\% class(x)
 },
  maxErrors = 10,
  sleep = 5
)
}
\arguments{
\item{expr}{A R expression using \code{\link[utils:download.file]{utils::download.file()}} to download a file.}

\item{isError}{A function to determine if a \emph{try-error} is returned}

\item{maxErrors}{An integer(1), specifying the maximal number of tries before
stopping retry.}

\item{sleep}{A numeric(1), specifying the time period in second to wait before
retry.}
}
\value{
0 if success, otherwise a \emph{try-error}.
}
\description{
Download files via HTTP with retry if failure occurs
}
\details{
This function is adopted from David Weber's post on the
StackOverflow with modification:
https://stackoverflow.com/questions/63340463/download-files-until-it-works.
}
\examples{
#' tmp_dir <- tempdir()
options(timeout = max(3000, getOption("timeout")))
## download feaureCounts results
count_url <- paste0(
    "https://zenodo.org/records/11458839/files/",
    "read_count_summary.RData?download=1"
)
retry_download({download.file(
    url = count_url,
    destfile = file.path(tmp_dir, "read_count_summary.RData"),
    mode = "wb"
)})
}
