% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_BSgenome.R
\name{UCSC2Ensembl}
\alias{UCSC2Ensembl}
\title{Convert a BSgenome from the UCSC to Ensembl style}
\usage{
UCSC2Ensembl(
  UCSC_BSgenome = NULL,
  genome_version = "GRCh38",
  seqname_alias = data.frame(ucsc = paste0("chr", c(seq_len(22), "X", "Y", "M")), ensembl
    = c(seq_len(22), "X", "Y", "MT"))
)
}
\arguments{
\item{UCSC_BSgenome}{An object of the \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} in the UCSC
style, such as BSgenome.Hsapiens.UCSC.hg38.}

\item{genome_version}{A character(1), specifying the genome version, such as
"GRCh38". Caution: make sure the genome_version here is the same as that
for the \code{\link[=make_ensdb]{make_ensdb()}}.}

\item{seqname_alias}{A data frame or a tab-delimited file to a data frame,
with two columns: ucsc and ensembl for UCSC-style seqnames and Ensembl-
style seqnames, respectively.}
}
\value{
An object of the \link[BSgenome:BSgenome-class]{BSgenome::BSgenome} in the Ensembl style,
such as BSgenome.Dvirilis.Ensembl.dvircaf1.
}
\description{
Convert a BSgenome from the UCSC style to the Ensembl style, changing
"chrM" to "MT", removing the "chr" prefix from chromosome/scaffold seqnames,
only keeping seqnames in the primary genome assembly, and changing the genome
version name form the UCSC name to the Ensembl name, such as from "hg38" to
"GRCh38".
}
\examples{
library("BSgenome.Hsapiens.UCSC.hg38")
ucsc_BSgenome <- BSgenome.Hsapiens.UCSC.hg38
ucsc_seqnames <-
    seqnames(ucsc_BSgenome)[!grepl("_alt|fix|hap\\\\d+",
                            seqnames(ucsc_BSgenome))]

## Wierd: BSgenome.Hsapiens.UCSC.hg19 has both chrM and chrMT for
## mitochondrial genome. renomve chrMT.

if (all(c("chrM", "chrMT") \%in\% ucsc_seqnames)) {
    ucsc_seqnames <- ucsc_seqnames[!ucsc_seqnames \%in\% "chrMT"]
}

ensembl_seqnames <- gsub(
    "^chr", "",
    gsub(
        "chrM$", "MT",
        gsub(
            "v", ".",
            gsub("_random|chr[^_]+_", "", ucsc_seqnames)
        )
    )
)
## for BSgenome.Hsapiens.UCSC.hg19, scaffold seqnames start with lower case,
## should be changed to upper case. For example, change "gl000231" to
## "GL000231". Add a suffix ".1" to these scaffold seqnames.

seqname_alias <- data.frame(ucsc = ucsc_seqnames,
                            ensembl = ensembl_seqnames)

bsgenome <- UCSC2Ensembl(
    UCSC_BSgenome = ucsc_BSgenome,
    genome_version = "GRCh38",
    seqname_alias = seqname_alias
)

}
