% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DNA_contamination.R
\name{check_read_assignment_stat}
\alias{check_read_assignment_stat}
\title{Visualize assignment statistics of reads/fragments by featureCounts}
\usage{
check_read_assignment_stat(assignment_stat = NULL)
}
\arguments{
\item{assignment_stat}{A data frame containing the assignment statistics with
GTF as annotation by the \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}} function, such as the
\code{stat} element in the sublist \code{gtf} in the output of the
\code{\link[=summarize_reads]{summarize_reads()}} function. You can access it via
\code{featureCounts()$gtf$stat}.}
}
\value{
A ggplot object, showing percentages and number of fragments in each
assignment category as determined by \code{\link[Rsubread:featureCounts]{Rsubread::featureCounts()}}.
}
\description{
Visualize assignment statistics of reads/fragments by featureCounts
}
\examples{
tmp_dir <- tempdir()
options(timeout = max(3000, getOption("timeout")))
## download feaureCounts results
count_url <- paste0(
    "https://zenodo.org/records/11458839/files/",
    "read_count_summary.RData?download=1"
)
download.file(
    url = count_url,
    destfile = file.path(tmp_dir, "read_count_summary.RData"),
    mode = "wb"
)
load(file.path(tmp_dir, "read_count_summary.RData"))

p <- check_read_assignment_stat(
    assignment_stat =
        counts_summary$gtf$stat
)
p

}
