% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_DNA_contamination.R
\name{check_sample_correlation}
\alias{check_sample_correlation}
\title{Check sample correlation}
\usage{
check_sample_correlation(counts = NULL)
}
\arguments{
\item{counts}{A numeric matrix or data frame containing gene expression count
data from an RNA-seq experiment, with row names for gene ID and column
names for sample names. If values are not integers, it will be rounded.}
}
\value{
NULL. A plot with pairwise scatter plots and Pearson correlation
coefficients is generated. When the sample size is big, save it as tiff
file.
}
\description{
Generate a panel of plots based on a count table, with the diagonal
showing the sample names, the lower triangle showing smoothed scatterplots
for gene expression of pairwise samples, and the upper triangle showing
Pearson correlation coefficients of gene expression of pairwise samples.
}
\examples{
tmp_dir <- tempdir()
options(timeout = max(3000, getOption("timeout")))
## download feaureCounts results
count_url <- paste0(
    "https://zenodo.org/records/11458839/files/",
    "read_count_summary.RData?download=1"
)
download.file(
    url = count_url,
    destfile = file.path(tmp_dir, "read_count_summary.RData"),
    mode = "wb"
)
load(file.path(tmp_dir, "read_count_summary.RData"))

check_sample_correlation(counts = counts_summary$gtf$counts)

}
